协议调试成功

+++ game-common/src/main/java/com/chen/config/MsgCode.java
@@ -2,9 +2,9 @@ package com.chen.config;
 
 public class MsgCode {
 
-    public static final int LoginAction = 1001;//登录
+    public static final short LoginAction = 1001;//登录
 
-    public static final int CreateRole = 1002;//创建角色
+    public static final short CreateRole = 1002;//创建角色
 
 
 }
